requirements.txt
#data-generator
pyyaml
psycopg2-binary

#CDC
python-dotenv
requests

#Consumer
kafka-python
google-cloud-storage
pyarrow
pandas

#dbt
dbt-snowflake

#Snowflake
snowflake-connector-python
